backend/requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
python-dotenv
loguru
requests

# AI
google-generativeai

# Embeddings
sentence-transformers
faiss-cpu